# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Abroad Matrimony API
  version: 1.0.0
  description: |
    REST API for the Abroad Matrimony matchmaking platform.

    **Authentication:** All user endpoints (except OTP request/verify and token refresh)
    require a Bearer access token issued by `POST /api/v1/auth/otp/verify` or
    `POST /api/v1/auth/token/refresh`.

    Admin endpoints require a Bearer admin token from `POST /admin/auth/login`.

    Webhook endpoints (`/stripe/webhook`, `/razorpay/webhook`) use provider signature
    verification instead of Bearer tokens.

    **Response envelope:** Every response follows `ApiResponse<T>`:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "hasMore": false, "cursor": null },
      "requestId": "req_abc123"
    }
    ```
    Errors add `"error": { "code": "ERROR_CODE", "message": "Human-readable message" }`.

    **All money values** are integers in paise (INR) or cents (USD). Never floats.

  contact:
    name: Abroad Matrimony Engineering
  license:
    name: Private
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.abroadmatrimony.com
    description: Production

# ── Security Schemes ──────────────────────────────────────────────────────────
components:
  securitySchemes:
    userBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /api/v1/auth/otp/verify or /token/refresh. Expires in 15 minutes.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT issued by /admin/auth/login. Longer-lived; carries AdminRole claim.

# ── Reusable Schemas ──────────────────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: "Phone must be in E.164 format" }
        details: { type: object, additionalProperties: true }
...
    HabitKey:
      type: string
      enum: [MORNING_ROUTINE, EXERCISE, HEALTHY_EATING, MINDFULNESS, JOURNALING, SOCIAL_CONNECTION, LEARNING]
      description: Canonical key for a trackable daily habit.
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:   { type: boolean, enum: [false] }
        error:     { $ref: '#/components/schemas/ApiError' }
        requestId: { type: string }

  # ── Reusable Parameters ───────────────────────────────────────────────────
  parameters:
...
    habitKeyPath:
      name: habitKey
      in: path
      required: true
      schema: { $ref: '#/components/schemas/HabitKey' }
      description: Habit canonical key

  # ── Reusable Responses ────────────────────────────────────────────────────
  responses:
    400BadRequest:
      description: Validation error or bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: VALIDATION_ERROR, message: "Phone must be in E.164 format" }
    401Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: UNAUTHORIZED, message: "Access token is invalid or expired" }
...
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    429TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: RATE_LIMIT_EXCEEDED, message: "Too many requests. Try again in 3600 seconds.", details: { retryAfterSeconds: 3600 } }
    500InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

# ── Tags ──────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/habits/{habitKey}/log:
    post:
      tags: [Habits]
      summary: Log a habit for today
      description: Idempotent daily log. Updates streak and consistency score dimension.
      operationId: logHabit
      security:
        - userBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/habitKeyPath'
      responses:
        '200':
          description: Habit logged
          content:
            application/json:
              example:
                success: true
                data: { habitKey: "MORNING_ROUTINE", loggedAt: "2026-05-29T07:30:00.000Z", currentStreak: 5, isNewLog: true }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '429': { $ref: '#/components/responses/429TooManyRequests' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/habits/streak:
    get:
      tags: [Habits]
      summary: Get overall habit streak summary
      operationId: getHabitStreak
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: Streak summary
          content:
            application/json:
              example:
                success: true
                data:
                  totalDaysLogged: 42
                  longestStreak: 14
                  currentStreak: 5
                  lastLoggedAt: "2026-05-29T07:30:00.000Z"
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/habits/reflection:
    get:
      tags: [Habits]
      summary: Get weekly rule-based reflection (HABIT-005)
      description: |
        Returns a personalised, rule-based weekly insight based on the user's habit logs
        over the last 30 days. Identifies most-consistent day-of-week and top streak habit.
        Result is cached in Redis for 7 days per user.
      operationId: getWeeklyReflection
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: Weekly reflection
          content:
            application/json:
              example:
                success: true
                data:
                  insight: "You are most consistent on Mondays. Your strongest habit is Drink water."
                  whyItMatters: "Partners who share consistent self-improvement habits show higher long-term compatibility."
                  weekStartDate: "2026-05-27"
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }
    post:
      tags: [Habits]
      summary: Add reflection note to today's habit log (existing endpoint)
      description: Adds or updates a personal reflection text on today's log entry for a specific habit.
      operationId: addHabitReflectionNote
      security:
        - userBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [habitKey, reflection]
              properties:
                habitKey: { type: string, enum: [HYDRATION, SLEEP, EXERCISE, MINDFULNESS, PRAYER, GRATITUDE, READING, JOURNALING, NO_ALCOHOL, NO_SMOKING] }
                reflection: { type: string, minLength: 1, maxLength: 1000 }
      responses:
        '200':
          description: Reflection saved
          content:
            application/json:
              example:
                success: true
                data: { habitKey: HYDRATION, logDate: "2026-05-29", reflection: "Felt refreshed" }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/habits/streaks:
    get:
      tags: [Habits]
      summary: Get all habits with weekly dots (HABIT-004)
      description: |
        Returns all 10 preset habits enriched with currentStreak, longestStreak, totalCompleted,
        and thisWeekDots — a 7-element boolean array (index 0=Monday … index 6=Sunday) indicating
        which days of the current week the habit was logged.
      operationId: getAllHabitsWithStreaks
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: All habits with streak data and weekly dots
          content:
            application/json:
              example:
                success: true
                data:
                  - habitKey: HYDRATION
                    label: Drink water
                    currentStreak: 3
                    longestStreak: 7
                    totalCompleted: 15
                    lastLoggedDate: "2026-05-29"
                    thisWeekDots: [true, true, false, true, false, false, false]
                meta: { total: 10 }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }
...
  /api/v1/habits/{habitKey}/history:
    get:
      tags: [Habits]
      summary: Per-habit weekly history for chart (HABIT-007)
      description: |
        Returns weekly completion data for a single habit over N weeks (default 8, max 52).
        Ordered oldest-first. Each entry has `dailyDots` (7-element boolean Mon–Sun array)
        so the client can render a calendar heatmap without extra computation.
      operationId: getHabitHistory
      parameters:
        - in: path
          name: habitKey
          required: true
          schema:
            type: string
            enum: [HYDRATION, SLEEP, EXERCISE, MINDFULNESS, PRAYER, GRATITUDE, READING, JOURNALING, NO_ALCOHOL, NO_SMOKING]
        - in: query
          name: weeks
          schema: { type: integer, minimum: 1, maximum: 52, default: 8 }
          description: Number of past weeks to return (default 8)
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: Weekly habit history
          content:
            application/json:
              example:
                success: true
                data:
                  - weekStartDate: "2026-04-07"
                    completedDays: 5
                    dailyDots: [true, true, false, true, true, false, false]
                  - weekStartDate: "2026-04-14"
                    completedDays: 7
                    dailyDots: [true, true, true, true, true, true, true]
                meta: { total: 8 }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }
...